requirements.txt
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
PyMuPDF==1.24.11
pytesseract==0.3.13
